change arbitrary module structure
