+++ zebra-state/src/service/arbitrary.rs
@@ -5,7 +5,12 @@ use proptest::{
 };
 use std::sync::Arc;
 
-use zebra_chain::{block::Block, fmt::SummaryDebug, parameters::NetworkUpgrade::Nu5, LedgerState};
+use zebra_chain::{
+    block::{Block, Height},
+    fmt::SummaryDebug,
+    parameters::NetworkUpgrade,
+    LedgerState,
+};
 use zebra_test::prelude::*;
 
 use crate::tests::Prepare;
@@ -55,7 +60,7 @@ impl Strategy for PreparedChain {
         let mut chain = self.chain.lock().unwrap();
         if chain.is_none() {
             // TODO: use the latest network upgrade (#1974)
-            let ledger_strategy = LedgerState::genesis_strategy(Nu5, None, false);
+            let ledger_strategy = LedgerState::genesis_strategy(NetworkUpgrade::Nu5, None, false);
 
             let (network, blocks) = ledger_strategy
                 .prop_flat_map(|ledger| {
@@ -85,4 +90,45 @@ impl Strategy for PreparedChain {
             network: chain.0,
         })
     }
+}
+
+/// Generate a chain to test legacy chain check.  
+pub(crate) fn partial_nu5_chain_strategy(
+    transaction_version_override: u32,
+    transaction_has_valid_network_upgrade: bool,
+    blocks_after_nu_activation: u32,
+) -> impl Strategy<
+    Value = (
+        Network,
+        Height,
+        zebra_chain::fmt::SummaryDebug<Vec<Arc<Block>>>,
+    ),
+> {
+    (any::<Network>(), NetworkUpgrade::branch_id_strategy()).prop_flat_map(
+        move |(network, random_nu)| {
+            // We are going to test the legacy chain with Canopy instead as we don't have a
+            // Nu5 activation height yet.
+            // TODO: update this to Nu5 after we have a height #1841
+            let mut nu = NetworkUpgrade::Canopy;
+            let nu_activation = NetworkUpgrade::Canopy.activation_height(network).unwrap();
+            let height = Height(nu_activation.0 + blocks_after_nu_activation);
+
+            // The `network_upgrade_override` will not be enough as when it is `None`,
+            // current network upgrade will be used (`NetworkUpgrade::Canopy`) which will be valid.
+            if !transaction_has_valid_network_upgrade {
+                nu = random_nu;
+            }
+
+            zebra_chain::block::LedgerState::height_strategy(
+                height,
+                Some(nu),
+                Some(transaction_version_override),
+                transaction_has_valid_network_upgrade,
+            )
+            .prop_flat_map(move |init| {
+                Block::partial_chain_strategy(init, blocks_after_nu_activation as usize)
+            })
+            .prop_map(move |partial_chain| (network, nu_activation, partial_chain))
+        },
+    )
 }
