+++ zebra-state/src/service/non_finalized_state.rs
@@ -5,8 +5,6 @@
 mod chain;
 mod queued_blocks;
 
-#[cfg(any(test, feature = "proptest-impl"))]
-pub mod arbitrary;
 #[cfg(test)]
 mod tests;
