+++ zebra-state/src/service/non_finalized_state/tests/prop.rs
@@ -2,7 +2,7 @@ use std::env;
 
 use zebra_test::prelude::*;
 
-use crate::service::non_finalized_state::{arbitrary::PreparedChain, Chain};
+use crate::service::{arbitrary::PreparedChain, non_finalized_state::Chain};
 
 const DEFAULT_PARTIAL_CHAIN_PROPTEST_CASES: u32 = 32;
