+++ zebra-state/src/service/tests.rs
@@ -3,14 +3,12 @@ use std::{env, sync::Arc};
 use futures::stream::FuturesUnordered;
 use tower::{util::BoxService, Service, ServiceExt};
 use zebra_chain::{
-    block::{Block, Height},
-    parameters::{Network, NetworkUpgrade},
-    serialization::ZcashDeserializeInto,
-    transaction, transparent,
+    block::Block, parameters::Network, serialization::ZcashDeserializeInto, transaction,
+    transparent,
 };
 use zebra_test::{prelude::*, transcript::Transcript};
 
-use crate::{init, BoxError, Config, Request, Response, Utxo};
+use crate::{init, service::arbitrary, BoxError, Config, Request, Response, Utxo};
 
 const LAST_BLOCK_HEIGHT: u32 = 10;
 
@@ -199,7 +197,7 @@ proptest! {
 
     #[test]
     fn at_least_one_transaction_with_network_upgrade(
-        (network, nu_activation_height, chain) in partial_nu5_chain_strategy(5, true, BLOCKS_AFTER_NU5)
+        (network, nu_activation_height, chain) in arbitrary::partial_nu5_chain_strategy(5, true, BLOCKS_AFTER_NU5)
     ) {
         let response = crate::service::legacy_chain_check(nu_activation_height, chain.into_iter(), network)
             .map_err(|error| error.to_string());
@@ -209,7 +207,7 @@ proptest! {
 
     #[test]
     fn no_transactions_with_network_upgrade(
-        (network, nu_activation_height, chain) in partial_nu5_chain_strategy(4, true, BLOCKS_AFTER_NU5)
+        (network, nu_activation_height, chain) in arbitrary::partial_nu5_chain_strategy(4, true, BLOCKS_AFTER_NU5)
     ) {
         let response = crate::service::legacy_chain_check(nu_activation_height, chain.into_iter(), network)
             .map_err(|error| error.to_string());
@@ -222,7 +220,7 @@ proptest! {
 
     #[test]
     fn at_least_one_transactions_with_inconsistent_network_upgrade(
-        (network, nu_activation_height, chain) in partial_nu5_chain_strategy(5, false, BLOCKS_AFTER_NU5)
+        (network, nu_activation_height, chain) in arbitrary::partial_nu5_chain_strategy(5, false, BLOCKS_AFTER_NU5)
     ) {
         let response = crate::service::legacy_chain_check(nu_activation_height, chain.into_iter(), network)
             .map_err(|error| error.to_string());
@@ -232,46 +230,4 @@ proptest! {
             Err("inconsistent network upgrade found in transaction".into())
         );
     }
-}
-
-// Utility functions
-
-fn partial_nu5_chain_strategy(
-    transaction_version_override: u32,
-    transaction_has_valid_network_upgrade: bool,
-    blocks_after_nu_activation: u32,
-) -> impl Strategy<
-    Value = (
-        Network,
-        Height,
-        zebra_chain::fmt::SummaryDebug<Vec<Arc<Block>>>,
-    ),
-> {
-    (any::<Network>(), NetworkUpgrade::branch_id_strategy()).prop_flat_map(
-        move |(network, random_nu)| {
-            // We are going to test the legacy chain with Canopy instead as we don't have a
-            // Nu5 activation height yet.
-            // TODO: update this to Nu5 after we have a height #1841
-            let mut nu = NetworkUpgrade::Canopy;
-            let nu_activation = NetworkUpgrade::Canopy.activation_height(network).unwrap();
-            let height = Height(nu_activation.0 + blocks_after_nu_activation);
-
-            // The `network_upgrade_override` will not be enough as when it is `None`,
-            // current network upgrade will be used (`NetworkUpgrade::Canopy`) which will be valid.
-            if !transaction_has_valid_network_upgrade {
-                nu = random_nu;
-            }
-
-            zebra_chain::block::LedgerState::height_strategy(
-                height,
-                Some(nu),
-                Some(transaction_version_override),
-                transaction_has_valid_network_upgrade,
-            )
-            .prop_flat_map(move |init| {
-                Block::partial_chain_strategy(init, blocks_after_nu_activation as usize)
-            })
-            .prop_map(move |partial_chain| (network, nu_activation, partial_chain))
-        },
-    )
 }
